NFC: Extract chooseNonConflicting to EstimatorHelper
Note this change only involves moving, deleting,
and adding prefixes where appropriate.
